move driver add crosscheck to fmi-import, remove apache csv dependency, format numeric output, misc refactor
