--- fmu-driver/src/test/kotlin/no/ntnu/ihb/fmi4j/fmudriver/crosscheck/XcOptions.kt
@@ -1,50 +1,0 @@
-package no.ntnu.ihb.fmi4j.fmudriver.crosscheck
-
-
-data class XcOptions(
-        var startTime: Double = 0.0,
-        var stopTime: Double = 10.0,
-        var stepSize: Double = 1e-3,
-        var relTol: Double = 1e-3,
-        var absTol: Double = 1e-3
-) {
-
-    companion object {
-
-        fun parse(txt: String): XcOptions {
-
-            return XcOptions().apply {
-
-                txt.trim().split("\n").forEach { line ->
-
-                    val split = line.split(",")
-                    if (split.isNotEmpty()) {
-                        val (fst, snd) = split
-                        when (fst) {
-                            "StartTime" -> {
-                                startTime = snd.toDouble()
-                            }
-                            "StopTime" -> {
-                                stopTime = snd.toDouble()
-                            }
-                            "StepSize" -> {
-                                stepSize = snd.toDouble()
-                            }
-                            "RelTol" -> {
-                                relTol = snd.toDouble()
-                            }
-                            "AbsTol" -> {
-                                absTol = snd.toDouble()
-                            }
-                        }
-
-                    }
-
-                }
-
-            }
-
-        }
-
-    }
-}
